+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.5.0] / 2023-05-31
 ### Updated
 - Update `ricaun.Nuke` to `1.5.0`
+### Fixed
+- Fix `PathConstruction` to `Globbing`
+- Fix `FileSystemTasks` to `AbsolutePathExtensions`
+- Fix `ToolPathResolver` to `NuGetToolPathResolver`
+- Fix `InstallationFiles` when folder.
 
 ## [1.4.3] / 2023-05-06
 ### Updated